requirements.txt
Django==1.6.7
MySQL-python==1.2.5
Pillow==2.6.1
South==1.0.1
argparse==1.2.1
distribute==0.7.3
django-bootstrap-breadcrumbs==0.6.0
django-debug-toolbar==1.2.2
django-simple-captcha==0.4.2
django-tinymce==1.5.3
easy-thumbnails==2.2
six==1.8.0
sqlparse==0.1.13
wsgiref==0.1.2
